Change Pet getMobHP() to int32
Changes getMobHP() to int32 so that hp is accurately returned

+++ src/map/lua/lua_petskill.cpp
@@ -111,7 +111,7 @@ float CLuaPetSkill::getTP()
     return static_cast<float>(m_PLuaPetSkill->getTP());
 }
 
-auto CLuaPetSkill::getMobHP() const -> uint8
+auto CLuaPetSkill::getMobHP() const -> int32
 {
     return m_PLuaPetSkill->getHP();
 }
